vision_service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
Pillow
pillow-heif
